Make 'Type' dropdown function when a new section is added.
The selector for finding the 'Type' dropdown is incorrect, and so the event handler to the document `change` event is not fired. Consequently, fields are not shown/hidden when a section is added inline with 'Add another content section'.

This commit fixes it by using a derivative of the same selector used elsewhere in the JS (`[class*=section_set]`).